trabajando en la biblioteca

+++ src/main/java/com/spootify/backend_spootify/Dtos/caratulaPlaylistDto.java
@@ -15,6 +15,6 @@ public class caratulaPlaylistDto {
     
     String Descripcion;
 
-    int cantidadCanciones;
+    BigDecimal cantidadCanciones;
 
 }
